metrics, moves more logic from logic.py into ga3.py. some major surgery to test_logic.py, moving large chunks into test_ga3.py
