+++ src/metrics/logic.py
@@ -3,14 +3,9 @@ from article_metrics.utils import ensure, lmap, create_or_update, first, ymd, lf
 from article_metrics.ga_metrics import core as ga_core
 from django.db.models import Sum, F
 from django.db.models.functions import TruncMonth
-from django.conf import settings
-from datetime import date, datetime
+from datetime import date
 from django.db import transaction
-import os
 import logging
-from urllib.parse import urlparse
-from functools import partial
-from collections import OrderedDict
 
 LOG = logging.getLogger(__name__)
 
@@ -50,80 +45,6 @@ def asmaps(rows):
     "convenience, converts a list of GA result rows into dicts"
     return [dict(zip(['identifier', 'date', 'views'], row)) for row in rows]
 
-def normalise_path(path):
-    return urlparse(path).path.lower()
-
-def parse_map_file(frame, contents=None):
-    contents and ensure(isinstance(contents, str), "'contents' must be a string'")
-
-    def _parse_line(line):
-        "the file is a simple 'cat nginx-redirect-file | grep prefix > outfile'"
-        line = line.strip()
-        if not line:
-            return
-        path, redirect = line.split("' '")
-        path, redirect = path.strip(" '"), redirect.strip(" ';")
-        prefix = frame['redirect-prefix']
-        ensure(redirect.startswith(prefix), "redirect doesn't start with redirect-prefix: %s" % line)
-        # /inside-elife/foobar => foobar
-        bits = redirect.strip('/').split('/', 1) # '/inside-elife/foobar' -> 'inside-elife/foobar' -> ['inside-elife, 'foobar']
-        redirect = models.LANDING_PAGE if len(bits) == 1 else bits[1]
-        return (path, redirect)
-    if contents:
-        contents = contents.splitlines()
-    else:
-        path = os.path.join(settings.GA_PTYPE_SCHEMA_PATH, frame['path-map-file'])
-        contents = open(path, 'r').readlines()
-    return OrderedDict(lfilter(None, lmap(_parse_line, contents)))
-
-#
-#
-#
-
-def process_prefixed_path(prefix, path):
-    path = normalise_path(path)
-    ensure(path.startswith(prefix), "path does not start with given prefix (%r): %s" % (prefix, path), ValueError)
-    # we could just dispense with the prefix and discard the first segment ...
-    prefix_len = len(prefix)
-    path = path[prefix_len:].strip().strip('/') # /events/foobar => foobar
-    identifier = path.split('/', 1)[0] # foobar/the-baz-in-bar-fooed-at-the-star => foobar
-    return identifier
-
-def process_mapped_path(mapping, path):
-    path = normalise_path(path)
-    return mapping.get(path)
-
-def generic_results_processor(ptype, frame, rows):
-
-    # TODO: will need to differentiate between ga3 and ga4 results
-
-    if 'path-map-file' in frame:
-        mapping = parse_map_file(frame)
-        path_processor = partial(process_mapped_path, mapping)
-    elif 'prefix' in frame:
-        path_processor = partial(process_prefixed_path, frame['prefix'])
-    elif 'path-map' in frame:
-        path_processor = partial(process_mapped_path, frame['path-map'])
-
-    ensure(path_processor, "generic results processing requires a 'prefix' or 'path-map' key.")
-
-    def _process(row):
-        try:
-            path, datestr, count = row
-            identifier = path_processor(path)
-            if identifier is None:
-                return # raise ValueError?
-            return {
-                'views': int(count),
-                'date': datetime.strptime(datestr, "%Y%m%d").date(),
-                'identifier': identifier,
-            }
-        except ValueError as err:
-            LOG.info("skipping row, bad value: %s" % str(err))
-        except BaseException as err:
-            LOG.exception("unhandled exception processing row: %s", str(err), extra={"row": row})
-    return list(filter(None, map(_process, rows)))
-
 #
 #
 #
@@ -144,27 +65,10 @@ def aggregate(normalised_rows):
     return rows
 
 def process_response(ptype, frame, response):
-
-    rows = response.get('rows')
-    if not rows:
-        LOG.warning("GA responded with no results", extra={'query': response['query'], 'ptype': ptype, 'frame': frame})
-        return []
-
-    # look for the "results_processor_frame_foo" function ...
-    #path = "metrics.{ptype}_type.results_processor_frame_{id}".format(ptype=ptype, id=frame['id'])
-
-    # ... and use the generic processor if not found.
-    #results_processor = load_fn(path) or generic_results_processor
-
-    # lsh@2023-02-13: nothing ever used the per-page-type results processor.
-    # instead it looks like `generic_results_processor` got fat attempting to handle everything.
-    results_processor = generic_results_processor
-
-    normalised = results_processor(ptype, frame, rows)
-
-    # todo: schema check normalised rows. should be easy
-
-    return normalised
+    era = ga_core.guess_era_from_response(response)
+    if era == ga_core.GA3:
+        return ga3.process_response(ptype, frame, response)
+    return ga4.process_response(ptype, frame, response)
 
 # ---
 
@@ -183,15 +87,9 @@ def build_ga_query__queries_for_frame(ptype, frame, start_date, end_date):
 # ---
 
 def query_ga(ptype, query, results_pp=MAX_GA_RESULTS, replace_cache_files=False):
-    ensure('end_date' in query, "query missing 'end_date' field")
-    end_date = query['end_date']
-    # urgh. todo?
-    if not is_date(end_date):
-        end_date = datetime.strptime(query['end_date'], '%Y-%m-%d').date()
-    era = ga_core.GA3 if end_date <= GA4_SWITCH_DATE else ga_core.GA4
-    if era == ga_core.GA3:
+    if ga_core.guess_era_from_query(query) == ga_core.GA3:
         return ga3.query_ga(ptype, query, results_pp, replace_cache_files)
-    return ga4.query_ga(ptype, query, results_pp, replace_cache_files)
+    return ga4.query_ga(ptype, query, None, replace_cache_files)
 
 # ---
